Add test utils to create unique urls for embed tests to fix
  'listen tcp 127.0.0.1:2380: bind: address already in use'

errors when running tests.

Address review comments:
- Add Apache copyright to the new file
- Fix typo Confg -> Config
- Move server/embed/testutil/url.go -> server/embed/util_test.go
- Add a new method to apply the test URLs to the config
- Make all config functions private
